+++ src/rle.h
@@ -4,4 +4,4 @@
 #include <list>
 
 
-std::pair<std::list<long>, std::list<long>> rle_encode(uint8_t *data, size_t size);
+std::pair<std::list<size_t>, std::list<uint8_t>> rle_encode(uint8_t *data, size_t size);
